requirements.txt
pandas
openai-whisper
fastapi
uvicorn[standard]
python-dotenv
openai-whisper
moviepy
keybert
torch
transformers
sentence-transformers
torchaudio
yake
accelerate
sentencepiece
faster-whisper